#1939: revert render refactoring
